🔨 [filestorage] Move function `fileexistspolicy` to `projects.generate`
